arrumando configurações do site
