+++ index.html
@@ -10,6 +10,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/App.jsx"></script>
-  </body>
+    <script type="module" src="./src/main.jsx"></script>
+  </body> 
 </html>
