+++ src/context/AuthContext.jsx
@@ -1,45 +1,41 @@
-import React, { createContext, useContext, useState, useEffect } from 'react';
-import { authApi } from '../api/Auth';
-import { useNavigate } from 'react-router-dom';
+import { createContext, useState, useEffect } from "react";
 
-const AuthContext = createContext(undefined);   
+export const AuthContext = createContext(undefined);   
 
-const AuthProvider = ({ children }) => {
-    const [user, setUser] = useState(() => {
+export const AuthProvider = ({ children }) => {
+    const [user, setUser] = useState(null);
+    const [token, setToken] = useState(null);
+
+    useEffect(() => {
         const storedUser = localStorage.getItem("user");
-        return storedUser ? JSON.parse(storedUser) : null;
-    });
-    
-    const [token, setToken] = useState(() => {
-        return localStorage.getItem("token") || null
-    });
+        const storedToken = localStorage.getItem("token");
+        
+        if (storedUser && storedToken) {
+            setUser(JSON.parse(storedUser));
+            setToken(storedToken);
+        }
+    }, []);
 
-    const login = (userData, authToken) => {
+    const login = (userData, tokenData) => {
         setUser(userData);
-        setToken(authToken);
+        setToken(tokenData);
+
         localStorage.setItem("user", JSON.stringify(userData));
-        localStorage.setItem("token", authToken);
+        localStorage.setItem("token", tokenData);
     };
-    
+
     const logout = () => {
         setUser(null);
         setToken(null);
-        localStorage.removeItem("token");
+
         localStorage.removeItem("user");
-        // navigate("/login")
-    }
-
-
-    useEffect(() => {
-    });
+        localStorage.removeItem("token");
+    };
 
     return (
         <AuthContext.Provider value={{ user, token, login, logout }}>
-            { children }
+            {children}
         </AuthContext.Provider>
-    )
-}
+    );
+};
 
-const useAuth = () => {
-    return useContext(AuthContext);
-}
