+++ src/main.jsx
@@ -1,0 +1,13 @@
+import React, { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+import { AuthProvider } from "@/context/AuthContext";
+import "@/utils/global.css";
+import App from "@/App";
+
+createRoot(document.getElementById("root")).render(
+  <StrictMode>
+    <AuthProvider>
+      <App />
+    </AuthProvider>
+  </StrictMode>,
+);
