+++ src/modals/Flavor/Flavor.module.css
