+++ src/modals/Flavor/components/SizeSelection.js
