+++ src/modals/Flavor/utils/getObservation.js
