+++ src/modals/Flavor/utils/isValid.js
