+++ src/modals/Select_pizza_type/select_pizza_type.css
